requirements.txt
marvin
python-dotenv
youtube-transcript-api